organized FavouriteRecipes and TaggedRecipes interface adapters into the display recipes package
