+++ src/interface_adapter/display_recipes/FavouriteRecipesController.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.display_recipes;
 
 import entity.User;
 import use_case.display_favourite_recipe.DisplayFavouriteInputBoundary;
@@ -11,7 +11,7 @@ public class FavouriteRecipesController {
     }
 
     public void execute(User user) {
-        DisplayFavouriteInputData displayFavouriteInputData = new DisplayFavouriteInputData(user);
+        DisplayFavouriteInputData displayFavouriteInputData = new DisplayFavouriteInputData(user.getUsername());
         displayFavouriteInteractor.execute(displayFavouriteInputData);
     }
 }
