--- src/interface_adapter/display_recipes/RecipesController.java
@@ -1,8 +1,0 @@
-package interface_adapter.display_recipes;
-
-public class RecipesController {
-
-    public RecipesController() {
-    }
-
-}
