--- src/interface_adapter/display_recipes/RecipesPresenter.java
@@ -1,7 +1,0 @@
-package interface_adapter.display_recipes;
-
-public class RecipesPresenter {
-
-        public RecipesPresenter() {
-        }
-}
